requirements.txt
Flask==3.0.3
rembg==2.0.58
pillow==10.4.0
python-dotenv==1.0.1
gunicorn==23.0.0
